+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 Check also [MIGRATION.md](MIGRATION.md) for possible compatibility problems.
 
+### 0.8.3:
+* Add `sliding` and `slidingToList` to `StreamEx`
+
 ### 0.8.2:
 * Add `groupByToEntry` and `partitionByToEntry`.
 * Add `collectAndThen` to `StreamEx` and `EntryStream`